[1.1] Auto downloading VVVVVV's environment
The plugin used to only download the libraries themselves, and packaged
data.zip, fonts and languages in the plugin itself. Not only does this
inflate the plugin size (making me unable to upload releases to GitHub)
but it also means the plugin had to be extracted before usage. Now that
the entire environment (data.zip, fonts/ and lang/) are all
automatically downloaded, these restrictions are no more.