1.4.0
This is a big one, and should've been multiple commits...

This commit introduces the SCREENSHOT button! You can either press F6,
or the "Screenshot" button in the top-right while you playtest.
In the past, you COULD press F6 and VVVVVV would attempt to take a
screenshot, but it wouldn't actually work because nothing is set up for
that. Now, pressing F6 tells the plugin itself to take a screenshot,
which gets saved to your Ved save directory. If you don't know where
that is, there's a button underneath the "Screenshot" button to open
the screenshots folder.

It is now possible to choose which playtesting mode you want through a
new checkbox in the "VVVVVV setup" screen. To easily tell whether
you're using the plugin or Ved's default playtesting, the play button
is now blue if the plugin is enabled.

If an error is detected, VLIB will no longer pretend to function until
you try to playtest -- it'll instead tell you the error, and then
disable itself. If you try to playtest in its disabled state, it'll ask
if you want to disable the plugin completely for this session and use
Ved's built-in playtesting instead.

Additionally, if the plugin notices that it loaded an unexpected
version of VLIB, it'll disable itself as well. Meaning, make sure to
keep VLIB updated! But if you don't, it won't crash anymore, and just
disable itself.

VLIB should now load level-specific assets for certain files instead of
just ignoring them completely. Although I have no idea who customizes
like... elephant.png. But, if you do, it works in the plugin now.

Regions are now supported! Before, using a region would just crash Ved
when you open the map, as the textures were not getting loaded. The
plugin now loads region graphics properly.

The bounds for sliders have been increased, so it should be easier to
mute your music/sounds properly.

Bugfixes:

- Quads are now cached instead of being created on-demand (slow!)
- One-way tiles are now tinted properly again after it broke in 1.3.0
- Screen flashes should no longer be translucent
- Keys will no longer get stuck down between playtest sessions